Refactor app blueprints
