+++ stockapp/alerts/__init__.py
@@ -1,0 +1,3 @@
+from .routes import alerts_bp
+
+__all__ = ['alerts_bp']
