+++ stockapp/portfolio/__init__.py
@@ -1,0 +1,3 @@
+from .routes import portfolio_bp
+
+__all__ = ['portfolio_bp']
