+++ templates/alerts.html
@@ -29,7 +29,7 @@
             </li>
             {% endfor %}
         </ul>
-        <a href="{{ url_for('watch.clear_alerts') }}" class="btn btn-sm btn-danger mt-3">Clear Alerts</a>
+        <a href="{{ url_for('alerts.clear_alerts') }}" class="btn btn-sm btn-danger mt-3">Clear Alerts</a>
         {% else %}
         <p>No alerts recorded.</p>
         {% endif %}
